+++ src/FileNotFoundException.php
@@ -1,7 +1,7 @@
 <?php
+declare(strict_types=1);
+
 /**
- * Pdf generator interface class
- *
  * @section LICENSE
  * This file is created by vianetz <info@vianetz.com>.
  * The code is distributed under the GPL license.
@@ -17,19 +17,8 @@
  * @license     http://www.gnu.org/licenses/gpl-3.0.txt GNU GENERAL PUBLIC LICENSE
  */
 
-namespace Vianetz\Pdf\Model;
+namespace Vianetz\Pdf;
 
-/**
- * Interface Vianetz_Pdf_Model_Generator_Interface
- */
-interface GeneratorInterface
+class FileNotFoundException extends \Exception
 {
-    /**
-     * Render the pdf document.
-     *
-     * @param DocumentInterface $documentModel
-     *
-     * @return string
-     */
-    public function renderPdfDocument(DocumentInterface $documentModel);
 }