+++ src/Model/Htmlable.php
@@ -17,8 +17,9 @@ declare(strict_types=1);
  * @license     http://www.gnu.org/licenses/gpl-3.0.txt GNU GENERAL PUBLIC LICENSE
  */
 
-namespace Vianetz\Pdf;
+namespace Vianetz\Pdf\Model;
 
-class Exception extends \Exception
+interface Htmlable
 {
+    public function toHtml(): string;
 }