+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [5.0.0] - 2025-01-10
+### Added
+- Merger for Zugferd PDFs, i.e. PDFs with XML attachments
+### Changed
+- TCPDF to Fpdf library for merging as default (both libraries supported now)
+- Several public interfaces to support type hints and make purpose clearer
+### Removed
+- Deprecated ZendPdf merger  
+
 ## [4.0.2] - 2024-10-13
 ### Changed
-- Updated dompdf library to 3.x
+- Dompdf library to version 3.x
 
 ## [4.0.1] - 2023-11-10
 ### Fixed